% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlpha.R
\name{add.alpha}
\alias{add.alpha}
\title{add.alpha function}
\usage{
add.alpha(col, alpha = 0.5)
}
\arguments{
\item{col}{The colour value(s)}

\item{alpha}{The transparency value. A number between 0 and 1 where 0 is fully transparent and 1 fully opaque. Defaults to 0.5}
}
\value{
The new colour value(s) with  transparency
}
\description{
This function adds a transparency value to colours in R.
}
\examples{
add.alpha("gray75", .25)
}
\keyword{alpha}
